requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0
psycopg2-binary>=2.9.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
praw>=7.7.0  # Reddit API
snscrape>=0.7.0  # Twitter scraping (no API key needed)
selenium>=4.15.0  # For JavaScript-heavy sites
playwright>=1.40.0  # Alternative to Selenium

# NLP & ML
transformers>=4.35.0
sentence-transformers>=2.2.0
torch>=2.1.0
nltk>=3.8.1
vaderSentiment>=3.3.2
spacy>=3.7.0

# OpenAI API (for search term generation)
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
joblib>=1.3.0

# Data visualization (optional)
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0

# Development
jupyter>=1.0.0
ipykernel>=6.27.0
pytest>=7.4.0
black>=23.12.0
flake8>=6.1.0
